handle UpdateDataEvents in detail, look at the josm event
major changes
- EditorObjectsProducer.java process() and combine()
- DataEvent.java and subclasses getJosmEvent() and getJosmData()

Prior to this commit, in filter util all primitives of the active edit
dataset in the LayerManager were tested against the layerMatchers for
each kendzi layer. This behavior is retained only when processing
NewDataEvent.

For UpdateDataEvent the primitives josm sent along with the change event
are now used to determine new / update / remove ids.

In NodeMovedEvent, josm sends only the node, but the 3d plugin needs to
know if a way, or more indirectly a relation is referencing this node.
For this reason, UpdateDataEvent.getData() uses expensive calls to
getReferrers().  It is desirable to further optimize this, e.g. by using
a HashMap<Node, List<EditableObject>> or similar.

In the UpdateDataEvent case, the filter util now uses an event-dependent
data collection for each layer that the 3d plugin needs to update.

This means the layerMatchers will have much less work, but, depending on
the specific event processed, setting up the data collection may lever-
age that gain somewhat. See specifically UpdateDataEvent.getJosmData()
that builds the data collection in response to NodeMovedEvent.